chore: Further refactoring
