+++ ra2ce/analyses/analysis_config_wrapper/analysis_config_wrapper_with_network.py
@@ -23,6 +23,7 @@
 from __future__ import annotations
 
 from pathlib import Path
+from ra2ce.analyses.analysis_config_data.analysis_config_data_validator_with_network import AnalysisConfigDataValidatorWithNetwork
 
 from ra2ce.analyses.analysis_config_wrapper.analysis_config_wrapper_base import (
     AnalysisConfigWrapperBase,
@@ -97,4 +98,5 @@ class AnalysisConfigWrapperWithNetwork(AnalysisConfigWrapperBase):
 
     def is_valid(self) -> bool:
         _file_is_valid = self.ini_file.is_file() and self.ini_file.suffix == ".ini"
-        return _file_is_valid and self.config_data.is_valid()
+        _validation_report = AnalysisConfigDataValidatorWithNetwork(self.config_data).validate()
+        return _file_is_valid and _validation_report.is_valid()
