+++ ra2ce/analyses/analysis_config_wrapper/analysis_config_wrapper_without_network.py
@@ -24,6 +24,7 @@ from __future__ import annotations
 
 import logging
 from pathlib import Path
+from ra2ce.analyses.analysis_config_data.analysis_config_data_validator_without_network import AnalysisConfigDataValidatorWithoutNetwork
 
 from ra2ce.analyses.analysis_config_wrapper.analysis_config_wrapper_base import (
     AnalysisConfigWrapperBase,
@@ -76,4 +77,5 @@ class AnalysisConfigWrapperWithoutNetwork(AnalysisConfigWrapperBase):
 
     def is_valid(self) -> bool:
         _file_is_valid = self.ini_file.is_file() and self.ini_file.suffix == ".ini"
-        return _file_is_valid and self.config_data.is_valid()
+        _validation_report = AnalysisConfigDataValidatorWithoutNetwork(self.config_data).validate()
+        return _file_is_valid and _validation_report.is_valid()
